fix: tariffid cant be edited/created. Backend now follows RBAC too

+++ tariff-backend/src/main/java/com/tariff/tariff_backend/controller/DashboardController.java
@@ -3,6 +3,7 @@ package com.tariff.tariff_backend.controller;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -11,15 +12,16 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.tariff.tariff_backend.model.Tariff;
 import com.tariff.tariff_backend.model.dashboard.DashboardMetrics;
 import com.tariff.tariff_backend.model.dashboard.DashboardResponse;
 import com.tariff.tariff_backend.model.dashboard.TariffPatchDTO;
 import com.tariff.tariff_backend.service.DashboardService;
+import com.tariff.tariff_backend.service.JwtService;
 
 @RestController
 @RequestMapping("/api/v1/dashboard")
@@ -27,13 +29,18 @@ import com.tariff.tariff_backend.service.DashboardService;
 public class DashboardController {
 
     private final DashboardService dashboardService;
+    private JwtService jwtService;
 
-    public DashboardController(DashboardService dashboardService) {
+    public DashboardController(DashboardService dashboardService, JwtService jwtService) {
         this.dashboardService = dashboardService;
+        this.jwtService = jwtService;
     }
 
     @PostMapping("/tariffs")
-    public ResponseEntity<?> createTariff(@RequestBody Tariff request) {
+    public ResponseEntity<?> createTariff(@RequestHeader("Authorization") String authHeader, @RequestBody TariffPatchDTO request) {
+        if (authHeader == null || !jwtService.hasRole(jwtService.getTokenFromHeader(authHeader), "admin")) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("You do not have enough permissions.");
+        }
         DashboardResponse dResponse = dashboardService.createTariff(request);
         if (!dResponse.getSuccess()) {
             return ResponseEntity.badRequest().body(dResponse);
@@ -42,7 +49,10 @@ public class DashboardController {
     }
 
     @PatchMapping("/tariffs/{tariffid}")
-    public ResponseEntity<?> updateTariff(@PathVariable Integer tariffid, @RequestBody TariffPatchDTO patchDTO) {
+    public ResponseEntity<?> updateTariff(@RequestHeader("Authorization") String authHeader, @PathVariable Integer tariffid, @RequestBody TariffPatchDTO patchDTO) {
+        if (authHeader == null || !jwtService.hasRole(jwtService.getTokenFromHeader(authHeader), "admin")) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("You do not have enough permissions.");
+        }
         DashboardResponse dResponse = dashboardService.updateTariff(tariffid, patchDTO);
         if (!dResponse.getSuccess()) {
             return ResponseEntity.badRequest().body(dResponse);
@@ -51,7 +61,10 @@ public class DashboardController {
     }
 
     @DeleteMapping("/tariffs/{tariffid}")
-    public ResponseEntity<?> deleteTariff(@PathVariable Integer tariffid) {
+    public ResponseEntity<?> deleteTariff(@RequestHeader("Authorization") String authHeader, @PathVariable Integer tariffid) {
+        if (authHeader == null || !jwtService.hasRole(jwtService.getTokenFromHeader(authHeader), "admin")) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("You do not have enough permissions.");
+        }
         DashboardResponse dResponse = dashboardService.deleteTariff(tariffid);
         if (!dResponse.getSuccess()) {
             return ResponseEntity.badRequest().body(dResponse);
